online-jobs-portal/backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
PyPDF2
python-docx
scikit-learn
numpy
Pillow
